Move ReverseDiff/Mooncake to integration tests, drop DI from Bijectors deps
ReverseDiff and Mooncake now have dedicated integration suites under
test/integration/{reversediff,mooncake}, mirroring the existing Enzyme setup.
Each backend gets its own CI workflow with blocking failures (Enzyme stays
non-blocking). test/ad/ disappears: the four backend-exercise wrappers fold
into the integration mains, and the rule-correctness test moves to
test/bijectors/chainrules.jl.

DifferentiationInterface drops out of Bijectors's [deps]/[compat]. The
DI-dependent test machinery (linked_optic_vec AD check, AD-vs-analytical
logjac, test_ad) moves from src/vector/test_utils.jl into the test resources;
VectorBijectors.test_all keeps only the structural checks.

Test cases now carry a tag::Symbol; new aggregate entry points
generate_ad_testcases() and generate_vector_testcases() return flat lists.
Each integration main.jl becomes one adtype list + one is_broken(c) predicate
+ two short loops. Enzyme collapses to a single runtime_activity+Const adtype
list across all cases.

The vector test_all coverage that lived in test/vector/*.jl is now driven
through test/vector/main.jl (machinery + GROUP-guarded loop) with per-tag
generators kept in the per-category sibling files.
